[MODI-168] TEST: 파티 컨트롤러 테스트 패키지 분리
